group category actions and reducers in separate folders
