+++ EventsExpress/ClientApp/src/actions/category/add-category.js
@@ -1,4 +1,4 @@
-﻿import { CategoryService } from '../services';
+﻿import { CategoryService } from '../../services';
 import get_categories from './category-list';
