+++ EventsExpress/ClientApp/src/components/Route guard/401.js
@@ -12,7 +12,7 @@ import { setAvatarError } from '../../actions/EditProfile/change-avatar'
 import { setEventError } from '../../actions/add-event'
 import { getCityError } from '../../actions/cities'
 import { setCountryError } from '../../actions/countries'
-import { setCategoryError } from '../../actions/add-category'
+import { setCategoryError } from '../../actions/category/add-category'
 import { getUsersError } from '../../actions/users'
 import { getEventError } from '../../actions/event-item-view'
 import { set2CommentError } from '../../actions/add-comment'
