+++ EventsExpress/ClientApp/src/containers/categories/SelectCategories.js
@@ -1,8 +1,8 @@
 ﻿import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import SelectCategories from '../components/SelectCategories/SelectCategories';
-import setUserCategory from '../actions/EditProfile/addUserCategory';
-import get_categories from '../actions/category-list';
+import SelectCategories from '../../components/SelectCategories/SelectCategories';
+import setUserCategory from '../../actions/EditProfile/addUserCategory';
+import get_categories from '../../actions/category/category-list';
 
 class SelectCategoriesWrapper extends Component {
     constructor(props) {
