+++ EventsExpress/ClientApp/src/containers/categories/category-item.js
@@ -5,9 +5,9 @@ import IconButton from "@material-ui/core/IconButton";
 import CategoryItem from "../../components/category/category-item";
 import CategoryEdit from "../../components/category/category-edit";
 
-import { add_category } from "../../actions/add-category";
-import { delete_category } from "../../actions/delete-category";
-import { set_edited_category } from "../../actions/add-category";
+import { add_category } from "../../actions/category/add-category";
+import { delete_category } from "../../actions/category/delete-category";
+import { set_edited_category } from "../../actions/category/add-category";
 
 
 class CategoryItemWrapper extends Component {
