+++ EventsExpress/ClientApp/src/reducers/category/add-category.js
@@ -1,8 +1,8 @@
-﻿import initialState from '../store/initialState';
+﻿import initialState from '../../store/initialState';
 
 import {
     SET_CATEGORY_ERROR, SET_CATEGORY_PENDING, SET_CATEGORY_SUCCESS
-} from '../actions/add-category';
+} from '../../actions/category/add-category';
 
 export const reducer = (state = initialState.add_category, action) => {
