+++ EventsExpress/ClientApp/src/reducers/root.js
@@ -14,8 +14,8 @@ import * as CancelNextEventSchedule from './cancel-next-eventSchedule';
 import * as CancelAllEventSchedules from './cancel-all-eventSchedules';
 import * as Events from './event-list';
 import * as EventSchedules from './eventSchedules-list';
-import * as AddCategories from './add-category';
-import * as Categories from './category-list';
+import * as AddCategories from './category/add-category';
+import * as Categories from './category/category-list';
 import * as Countries from './countries';
 import * as Cities from './cities';
 import * as Users from './users';
